changed file location, and resolved most issues
