+++ manual_run/csv_weight_run_revised/attributes.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class Attributes(Enum):
+    SCORE = 0
+    TUTOR_PREFERENCE = 1
+    GROUP_SIZE = 2
+    TIMESLOT_AVAILABILITY = 3
