+++ manual_run/csv_weight_run_revised/weight_run.py
@@ -2,24 +2,25 @@ import csv
 import os
 from math import ceil
 from os import path
-from typing import List
 
 import typer
+from manual_run.csv_weight_run.attributes import Attributes
+from manual_run.csv_weight_run.data_provider import DataProvider
+from manual_run.csv_weight_run.default_scenario import DefaultScenario
+from manual_run.csv_weight_run.variables import Variables
 
 from api.ai.interfaces.algorithm_config import PriorityAlgorithmConfig
-from api.models.enums import AlgorithmType, DiversifyType, ScenarioAttribute
-from benchmarking.evaluations.goals import DiversityGoal, WeightGoal
-from benchmarking.evaluations.interfaces import Goal, Scenario
-from benchmarking.evaluations.metrics.cosine_similarity import AverageCosineDifference
-from benchmarking.runs.csv_weight_run_revised.attributes import Attributes
-from benchmarking.runs.csv_weight_run_revised.data_provider_revised import DataProvider
+from api.models.enums import AlgorithmType
+from benchmarking.evaluations.metrics.cosine_similarity import \
+    AverageCosineDifference
 from benchmarking.runs.interfaces import Run
-from benchmarking.simulation.insight import Insight
-from benchmarking.simulation.simulation_set import SimulationSet, SimulationSetArtifact
+from benchmarking.simulation.simulation_set import (SimulationSet,
+                                                    SimulationSetArtifact)
 from benchmarking.simulation.simulation_settings import SimulationSettings
 
 
 class WeightRun(Run):
+    
     def start(self, num_trials: int = 1, generate_graphs: bool = True):
         scenario = DefaultScenario()
 
@@ -31,19 +32,15 @@ class WeightRun(Run):
             ),
             AverageCosineDifference(
                 name="Timeslot Cosine Difference",
-                attribute_filter=[ScenarioAttribute.TIMESLOT_AVAILABILITY.value],
+                attribute_filter=[Attributes.TIMESLOT_AVAILABILITY.value],
             ),
         ]
         student_provider = DataProvider()
-        print(student_provider)
-        team_size = 4  # enter the desired team size
         artifact: SimulationSetArtifact = SimulationSet(
             settings=SimulationSettings(
-                num_teams=ceil(student_provider.num_students / team_size),
+                num_teams=ceil(student_provider.num_students / Variables.team_size),
                 scenario=scenario,
                 student_provider=student_provider,
-                # cache_key=f"weight_run_for_bowen/weight_run_2/",
-                cache_key=f"csv_weight_run_revised/weight_run_revised/",
             ),
             algorithm_set={
                 AlgorithmType.PRIORITY: [
@@ -55,22 +52,22 @@ class WeightRun(Run):
                     ),
                 ]
             },
-        ).run(num_runs=num_trials)
+        ).run(num_runs=1)
         team_set = list(artifact.values())[0][0][0]
 
-        insight_output_set = Insight.get_output_set(artifact, metrics)
-        print(insight_output_set)
-
-        # Enter the data fields from the provided CSV file
-        data_fields = [
-            ["ResponseId", "Q8", "Q4", "Q5", "zPos", "TeamSizeViolation", "TeamId"]
-        ]
+        self.create_csv(team_set, student_provider)
+        
+        
+    def create_csv(self, team_set, student_provider):
+         # Enter the data fields from the provided CSV file
+        Variables.data_fields[0].append("TeamId")
+        
         for team in team_set.teams:
             for student in team.students:
                 attributes = student.attributes
 
                 unique_id = student_provider.get_student(student.id)
-                time_slot = attributes[ScenarioAttribute.TIMESLOT_AVAILABILITY.value][0]
+                time_slot = attributes[Attributes.TIMESLOT_AVAILABILITY.value][0]
                 tutor_preference = attributes[Attributes.TUTOR_PREFERENCE.value][0]
                 group_size = attributes[Attributes.GROUP_SIZE.value][0]
 
@@ -96,7 +93,7 @@ class WeightRun(Run):
                     else ""
                 )
 
-                data_fields.append(
+                Variables.data_fields.append(
                     [
                         unique_id,
                         data_fields_input_1,
@@ -107,53 +104,18 @@ class WeightRun(Run):
                         team.id,
                     ]
                 )
-
         # Directory where the new CSV file will be written
         output_dir = path.join(path.dirname(__file__), "new_csv_output")
         if not os.path.exists(output_dir):
             os.makedirs(output_dir)
 
-        output_file = os.path.join(output_dir, "result.csv")
-
+        output_file = os.path.join(output_dir, Variables.output_file_name)
+        
         with open(output_file, "w+", newline="") as f:
             writer = csv.writer(f)
-            for row in data_fields:
+            for row in Variables.data_fields:
                 writer.writerow(row)
 
 
-class DefaultScenario(Scenario):
-    @property
-    def name(self) -> str:
-        return "Diversify on score and concentrate on timeslot"
-
-    @property
-    def goals(self) -> List[Goal]:
-        """
-        This function is meant to create a list of scenerio goals to be followed for team formation.
-
-        Returns:
-            List[Goal]: A list of goals for team formation.
-        """
-        return [
-            # Concentrate on student timeslot availability
-            DiversityGoal(
-                DiversifyType.CONCENTRATE,
-                ScenarioAttribute.TIMESLOT_AVAILABILITY.value,
-            ),
-            # Diversify students on tutor preference
-            DiversityGoal(
-                DiversifyType.DIVERSIFY,
-                Attributes.TUTOR_PREFERENCE.value,
-            ),
-            # Diversify students on score
-            DiversityGoal(
-                DiversifyType.DIVERSIFY,
-                Attributes.SCORE.value,
-            ),
-            WeightGoal(diversity_goal_weight=1),
-        ]
-
-
-# Run script from command line: python3 -m benchmarking.runs.csv_weight_run_revised.weight_run_revised
 if __name__ == "__main__":
     typer.run(WeightRun().start)
